Revert "Unexport FAILED status helpers and test in-package"
This reverts commit 13e1d346cdaf1cd985c6895fd79108225d700927.
